beginning of CesiumRoute View window imageryLayers control
some fixes for Time window

fixed HttpStatsReporter, which was listening on BusEvents when the
DataClient actor was already translating it to HttpContent

added a LineBufferArchiveReader as a convenient class for CVS
import/replay

added a FISSTArchiveReader as example

fixed overflow when parsing doubles in CharSeqByteSlice

added Sentinel and related SentinelRoute to show how to import Delphire
Sentinel sensor data

+++ race-core/src/main/scala/gov/nasa/race/common/CharSeqByteSlice.scala
@@ -310,7 +310,7 @@ trait CharSeqByteSlice extends ByteSlice with CharSequence {
     if (b == '.') {
       i += 1
       var m: Long = 1
-      var frac: Int = 0
+      var frac: Long = 0
       while (i < iMax && {b=data(i); isDigit(b)}){
         frac = (frac*10) + digitValue(b)
         m *= 10
